reorganized training module
